remove MailerInterface usage (fixes #225)
fixes #217 as well
