+++ backend/.env
@@ -48,9 +48,12 @@ SES_SECRET_ACCESS_KEY=
 SES_NEWSLETTER_CONFIGURATION_SET_NAME=newsletter
 ###< app ###
 
-MAILER_DSN=
 RELAY_URL=https://relay.hyvor.com
 RELAY_API_KEY=
+
+# Email to send system notifications from (e.g. domain verification mails, etc.)
+# Important: this domain must be verified in Relay, in the same project as the RELAY_API_KEY
+SYSTEM_MAIL_FROM=
 
 ###> symfony/lock ###
 # Choose one of the stores below
