+++ backend/src/Api/Local/TemplateController.php
@@ -10,12 +10,10 @@ use App\Service\Newsletter\NewsletterService;
 use App\Service\Template\HtmlTemplateRenderer;
 use App\Service\Template\TemplateService;
 use App\Service\Template\TemplateVariableService;
-use App\Service\UserInvite\EmailNotificationService;
 use App\Tests\Factory\NewsletterFactory;
 use App\Tests\Factory\SendFactory;
 use App\Tests\Factory\SubscriberFactory;
 use Doctrine\ORM\EntityManagerInterface;
-use Hyvor\Internal\Billing\BillingInterface;
 use Hyvor\Internal\Component\InstanceUrlResolver;
 use Hyvor\Internal\InternalConfig;
 use Hyvor\Internal\Internationalization\StringsFactory;
