+++ backend/src/Service/Subscriber/SubscriberCreatedListener.php
@@ -3,9 +3,10 @@
 namespace App\Service\Subscriber;
 
 use App\Entity\Type\SubscriberStatus;
-use App\Event\Subscriber\CreateSubscriberEvent;
 use App\Service\Content\ContentService;
+use App\Service\Integration\Relay\RelayApiClient;
 use App\Service\SendingProfile\SendingProfileService;
+use App\Service\Subscriber\Event\SubscriberCreatedEvent;
 use App\Service\Template\HtmlTemplateRenderer;
 use App\Service\Template\TemplateService;
 use App\Service\Template\TemplateVariableService;
@@ -13,12 +14,9 @@ use Hyvor\Internal\Component\InstanceUrlResolver;
 use Hyvor\Internal\InternalConfig;
 use Hyvor\Internal\Internationalization\StringsFactory;
 use Hyvor\Internal\Util\Crypt\Encryption;
-use Hyvor\Internal\Util\Transfer\Encryptable;
 use Symfony\Component\Clock\ClockAwareTrait;
 use Symfony\Component\EventDispatcher\Attribute\AsEventListener;
-use Symfony\Component\Mailer\MailerInterface;
 use Symfony\Component\Mime\Email;
-use Twig\Environment;
 
 #[AsEventListener]
 final class SubscriberCreatedListener
@@ -26,7 +24,6 @@ final class SubscriberCreatedListener
     use ClockAwareTrait;
 
     public function __construct(
-        private MailerInterface         $mailer,
         private SendingProfileService   $sendingProfileService,
         private Encryption              $encryption,
         private ContentService          $contentService,
@@ -36,11 +33,12 @@ final class SubscriberCreatedListener
         private readonly StringsFactory $stringsFactory,
         private InstanceUrlResolver     $instanceUrlResolver,
         private InternalConfig          $internalConfig,
+        private RelayApiClient          $relayApiClient,
     )
     {
     }
 
-    public function __invoke(CreateSubscriberEvent $event): void
+    public function __invoke(SubscriberCreatedEvent $event): void
     {
         $subscriber = $event->getSubscriber();
         $newsletter = $subscriber->getNewsletter();
@@ -118,6 +116,6 @@ final class SubscriberCreatedListener
             ->html($this->htmlTemplateRenderer->render($template, $variables))
             ->subject($heading . ' to ' . $newsletter->getName());
 
-        $this->mailer->send($email);
+        $this->relayApiClient->sendEmail($email);
     }
 }
