+++ backend/src/Service/Subscriber/SubscriberService.php
@@ -3,26 +3,23 @@
 namespace App\Service\Subscriber;
 
 use App\Entity\Media;
-use App\Entity\NewsletterList;
 use App\Entity\Newsletter;
+use App\Entity\NewsletterList;
 use App\Entity\Send;
 use App\Entity\Subscriber;
 use App\Entity\SubscriberExport;
 use App\Entity\Type\SubscriberExportStatus;
 use App\Entity\Type\SubscriberSource;
 use App\Entity\Type\SubscriberStatus;
-use App\Event\Subscriber\CreateSubscriberEvent;
 use App\Repository\SubscriberRepository;
 use App\Service\Subscriber\Dto\UpdateSubscriberDto;
+use App\Service\Subscriber\Event\SubscriberCreatedEvent;
 use App\Service\Subscriber\Message\ExportSubscribersMessage;
-use App\Service\UserInvite\EmailNotificationService;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Clock\ClockAwareTrait;
 use Symfony\Component\EventDispatcher\EventDispatcherInterface;
 use Symfony\Component\Messenger\MessageBusInterface;
-use Symfony\Component\String\Exception\InvalidArgumentException;
-use Symfony\Component\DependencyInjection\Attribute\Autoconfigure;
 
 class SubscriberService
 {
@@ -87,8 +84,8 @@ class SubscriberService
         $this->em->persist($subscriber);
         $this->em->flush();
 
-        $event = new CreateSubscriberEvent($subscriber);
-        $this->eventDispatcher->dispatch($event, CreateSubscriberEvent::class);
+        $event = new SubscriberCreatedEvent($subscriber);
+        $this->eventDispatcher->dispatch($event, SubscriberCreatedEvent::class);
 
         return $subscriber;
     }
@@ -205,8 +202,8 @@ class SubscriberService
         $this->em->persist($subscriber);
         $this->em->flush();
 
-        $event = new CreateSubscriberEvent($subscriber);
-        $this->eventDispatcher->dispatch($event, CreateSubscriberEvent::class);
+        $event = new SubscriberCreatedEvent($subscriber);
+        $this->eventDispatcher->dispatch($event, SubscriberCreatedEvent::class);
 
         return $subscriber;
     }
